Divide component to display multiple output form
